code tidying + cleaning, new imports, comment out draft code
Lots of little cleanup removing and adding spaces to code
to make the code tidier.
Also corrected a few typos.
Commented out functions that are drafted but not yet ready to run,
so that the package can be documented and built successfully.
Added more packages to imports based on existing functions.
Re-roxygenized the docs
Added docs for the datasets.